+++ src/main/java/com/enigma/restgarden/service/reservation/ReservationServiceDbImpl.java
@@ -125,7 +125,7 @@ public class ReservationServiceDbImpl implements ReservationService{
             Reservation reservation = new Reservation(grave, user, reservationDto.getTotalSlot(), reservationDto.getStatus(), reservationDto.getDescription() );
             reservation.setGravePrice(grave.getPrice());
             Double bookingFee = reservation.getGravePrice() * 0.2;
-            reservation.setTotalPayment((reservation.getGravePrice()+bookingFee)*reservation.getTotalSlot());
+            reservation.setTotalPayment((reservation.getGravePrice()+bookingFee.intValue())*reservation.getTotalSlot());
 
             reservation.setStatus("Reserved");
